app/requirements.txt
flask>=1.0
flask-sqlalchemy
flask-migrate
psycopg2
